Struct freebase1_sandbox::Freebase

pub struct Freebase<C, NC, A> {
    // some fields omitted
}

Central instance to access all Freebase related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate freebase1_sandbox;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use freebase1_sandbox::Freebase;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Freebase::new(hyper::Client::new(), auth);

Methods

impl<'a, C, NC, A> Freebase<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn new(client: C, authenticator: A) -> Freebase<C, NC, A>