Struct freebase1_sandbox::ReconcileGet

pub struct ReconcileGet {
    pub costs: Option<ReconcileGetCosts>,
    pub warning: Vec<ReconcileGetWarning>,
    pub candidate: Vec<ReconcileCandidate>,
    pub match_: Option<ReconcileCandidate>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

costs

Server costs for reconciling.

warning

If filled, then there were recoverable problems that affected the request. For example, some of the properties were ignored because they either are not valid Freebase predicates or are not indexed for reconciliation. The candidates returned should be considered valid results, with the caveat that sections of the request were ignored as specified by the warning text.

candidate

If filled, then the listed candidates are potential matches, and such should be evaluated by a more discerning algorithm or human. The matches are ordered by confidence.

match_

If filled, this entity is guaranteed to match at requested confidence probability (default 99%).